+++ src/wi_stuff.h
@@ -34,6 +34,8 @@ typedef enum
   ShowNextLoc
 } stateenum_t;
 
+extern int wi_more_widgets; // [Cherry]
+
 // Called by main loop, animate the intermission.
 void WI_Ticker (void);